added a few more commands, havent tested
