+++ bot/commands/__init__.py
@@ -102,114 +102,6 @@ def seconds_to_string(actual_seconds, extra_parts=1):
 def get_role_id(guild_id, rank_name):
 	return roles.get(str(guild_id), {}).get(rank_name)
 
-@betterbot.command(name='whois')
-async def whois(message, member: Member=None):
-	if not member:
-		return await message.send('Do `!whois @member` to get information on that user.')
-	data = await db.get_minecraft_data(member.id)
-	if not data:
-		return await message.send(embed=discord.Embed(
-			description="This user hasn't linked their account yet. Tell them to do **!link**."
-		))
-	print(data)
-	embed = discord.Embed(
-		title=f'Who is {member}'
-	)
-
-	uuid = data['uuid']
-
-	embed.add_field(
-		name='IGN',
-		value=data['ign'],
-		inline=False,
-	)
-	embed.add_field(
-		name='UUID',
-		value=uuid,
-		inline=False,
-	)
-
-	embed.set_thumbnail(url=f'https://crafatar.com/renders/head/{uuid}?overlay=1')
-
-	await message.channel.send(embed=embed)
-
-
-@betterbot.command(name='debugtime')
-async def debugtime(message, length: Time):
-	'Debugging command to test time'
-	await message.send(seconds_to_string(length))
-
-@betterbot.command(name='debugmember')
-async def debugmember(message, member: Member=None):
-	await message.send(embed=discord.Embed(
-		description=f'<@{member.id}>'
-	))
-
-@betterbot.command(name='mute', bot_channel=False)
-async def mute(message, member: Member, length: Time=0, reason: str=None):
-	'Mutes a member for a specified amount of time'
-
-	if not (
-		has_role(message.author.id, 717904501692170260, 'helper')
-		or has_role(message.author.id, 717904501692170260, 'trialhelper')
-	): return
-
-	if not member or not length:
-		return await message.channel.send(
-			'Invalid command usage. Example: **!mute gogourt 10 years nerd**'
-		)
-
-	if reason:
-		reason = reason.strip()
-
-	if reason:
-		mute_message = f'<@{member.id}> has been muted for "**{reason}**".'
-	else:
-		mute_message = f'<@{member.id}> has been muted.'
-
-	await message.send(embed=discord.Embed(
-		description=mute_message
-	))
-
-	await db.add_infraction(
-		member.id,
-		'mute',
-		reason
-	)
-
-	try:
-		await member.send(f'You were muted for "**{reason}**"')
-	except:
-		pass
-
-	try:
-		await mute_user(
-			member,
-			length,
-			message.guild.id if message.guild else None
-		)
-	except discord.errors.Forbidden:
-		await message.send("I don't have permission to do this")
-
-
-@betterbot.command(name='unmute', bot_channel=False)
-async def unmute(message, member: Member):
-	'Removes a mute from a member'
-
-	if not (
-		has_role(message.author.id, 717904501692170260, 'helper')
-		or has_role(message.author.id, 717904501692170260, 'trialhelper')
-	):
-		return
-
-	await unmute_user(
-		member.id,
-		reason=f'Unmuted by {str(message.author)}'
-	)
-
-	await message.send(embed=discord.Embed(
-		description=f'<@{member.id}> has been unmuted.'
-	))
 
 @betterbot.command(name='gulag')
 async def gulag(message, length: Time=60):
